+++ src/main/java/com/raoulvdberge/refinedstorage/apiimpl/storage/StorageCacheItem.java
@@ -11,11 +11,12 @@ import net.minecraft.item.ItemStack;
 
 import javax.annotation.Nonnull;
 import java.util.ArrayList;
+import java.util.Collections;
 import java.util.List;
 
 public class StorageCacheItem implements IStorageCache<ItemStack> {
     private INetworkMaster network;
-    private List<IStorage<ItemStack>> storages = new ArrayList<>();
+    private List<IStorage<ItemStack>> storages = Collections.synchronizedList(new ArrayList<>());
     private IStackList<ItemStack> list = API.instance().createItemStackList();
 
     public StorageCacheItem(INetworkMaster network) {